httpio: Clean up and fix httpio.Session
