--- gridstatus/httpio.py
@@ -1,1 +1,0 @@
-from gridstatus.httpio.httpio_pandas import HttpioPandas
